+++ src/ContainerRegistry/ContainerRegistry/Models/ContainerRegistryDataPlaneClient.cs
@@ -14,23 +14,36 @@
 
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
+using Microsoft.Azure.Commands.Common.Authentication.Authentication;
+using Microsoft.Azure.Commands.ContainerRegistry.Models;
+using Microsoft.Azure.Commands.ContainerRegistry.DataPlaneOperations;
 using Microsoft.Azure.ContainerRegistry;
 using Microsoft.Azure.Management.ContainerRegistry;
 using Microsoft.Rest;
 using System;
+using System.Collections.Generic;
+using Microsoft.WindowsAzure.Commands.Common;
+using Microsoft.Azure.Commands.Profile.Utilities;
+using System.Text.Json;
+using System.Linq;
 
 namespace Microsoft.Azure.Commands.ContainerRegistry
 {
     public class ContainerRegistryDataPlaneClient
     {
         private AzureContainerRegistryClient _client;
-        private ServiceClientCredentials _clientCredential;
-        private IAccessToken _accessToken;
+        private string _accessToken = default(string);
         private string _endPoint;
-        private const string _grantType = "access_token";
-        private const string _scope = "registry:catalog:*";
-        private const string _https = "https://";
+        private IAzureContext _context;
         private readonly string _suffix;
+        
+        private const string _grantType = "access_token";
+        private const string _scopeDefault = "registry:catalog:*";
+        private const string _https = "https://";
+        private const string _refreshTokenKey = "AcrRefreshToken";
+        private const string _acrTokenCacheKey = SharedComponentKeys.AcrTokenCacheKey;
+        private const int _minutesBeforeExpiration = 5;
+
 
         public Action<string> VerboseLogger { get; set; }
         public Action<string> ErrorLogger { get; set; }
@@ -38,22 +51,63 @@ namespace Microsoft.Azure.Commands.ContainerRegistry
 
         public ContainerRegistryDataPlaneClient(IAzureContext context)
         {
-            _clientCredential = AzureSession.Instance.AuthenticationFactory.GetServiceClientCredentials(context, AzureEnvironment.Endpoint.ResourceManager);
-            _accessToken = AzureSession.Instance.AuthenticationFactory.Authenticate(context.Account, context.Environment, context.Tenant.Id, null, ShowDialog.Never, null, context.Environment.GetTokenAudience(AzureEnvironment.Endpoint.ResourceManager));
+            _context = context;
             _suffix = context.Environment.ContainerRegistryEndpointSuffix;
-            _client = AzureSession.Instance.ClientFactory.CreateCustomArmClient<AzureContainerRegistryClient>(_clientCredential);
+            ServiceClientCredentials clientCredential = new RenewingTokenCredential(new ExternalAccessToken(_accessToken, () => _accessToken));
+            _client = AzureSession.Instance.ClientFactory.CreateCustomArmClient<AzureContainerRegistryClient>(clientCredential);
         }
 
-        public string GetRefreshToken()
+        public IDictionary<string, Tuple<string, DateTime>> TryGetTokenCache()
         {
-            var response = _client.RefreshTokens.GetFromExchangeAsync(grantType: _grantType, service: _endPoint, accessToken: _accessToken.AccessToken);
-            return response.GetAwaiter().GetResult().RefreshTokenProperty;
+            IDictionary<string, Tuple<string, DateTime>> cache;
+            if (!AzureSession.Instance.TryGetComponent(_acrTokenCacheKey, out cache))
+            {
+                AzureSession.Instance.RegisterComponent<IDictionary<string, Tuple<string, DateTime>>>(_acrTokenCacheKey, () => new Dictionary<string, Tuple<string, DateTime>>());
+                AzureSession.Instance.TryGetComponent(_acrTokenCacheKey, out cache);
+            }
+            return cache;
         }
 
-        public string GetAccessToken()
+        public AzureContainerRegistryClient GetClient()
         {
-            var response = _client.AccessTokens.GetAsync(service: _endPoint, scope: _scope, refreshToken: GetRefreshToken());
-            return response.GetAwaiter().GetResult().AccessTokenProperty;
+            return _client;
+        }
+
+        public string TryAuthenticate(string scope = _scopeDefault)
+        {
+            _accessToken = TryGetToken(scope);
+            return _accessToken;
+        }
+
+        private string TryGetToken(string key)
+        {
+            IDictionary<string, Tuple<string, DateTime>> cache = TryGetTokenCache();
+            Tuple<string, DateTime> value;
+            cache.TryGetValue(key, out value);
+            if (value == null || isTokenExpired(value))
+            {
+                string token = key.Equals(_refreshTokenKey) ? getRefreshToken() : getAccessToken(key);
+                value = new Tuple<string, DateTime>(token, getTokenExpiration(token));
+                cache[key] = value;
+            }
+            return value.Item1;
+        }
+
+        private DateTime getTokenExpiration(string token)
+        {
+            string decodedToken = Base64UrlHelper.DecodeToString(token.Split('.')[1]);
+            int unixTimeSeconds = JsonDocument.Parse(decodedToken)
+                                      .RootElement
+                                      .EnumerateObject()
+                                      .Where(p => p.Name == "exp")
+                                      .Select(p => p.Value.GetInt32())
+                                      .First();
+            return DateTimeOffset.FromUnixTimeSeconds(unixTimeSeconds).UtcDateTime;
+        }
+
+        private bool isTokenExpired(Tuple<string, DateTime> value)
+        {
+            return (value.Item2 - DateTime.UtcNow).Minutes <= _minutesBeforeExpiration;
         }
 
         public void SetEndPoint(string RegistryName)
@@ -66,10 +120,105 @@ namespace Microsoft.Azure.Commands.ContainerRegistry
         {
             return _endPoint;
         }
-
-        public Rest.Azure.AzureOperationResponse CheckRegistry()
+        
+        private string getAccessToken(string scope)
         {
-            return _client.V2Support.CheckWithHttpMessagesAsync().GetAwaiter().GetResult();
+            string refreshToken = TryGetToken(_refreshTokenKey);
+            return GetClient()
+                    .AccessTokens
+                    .GetAsync(service: _endPoint, scope: scope, refreshToken: refreshToken)
+                    .GetAwaiter()
+                    .GetResult()
+                    .AccessTokenProperty;
+        }
+
+        private string getRefreshToken()
+        {
+            string armAccessToken = AzureSession
+                                    .Instance.AuthenticationFactory
+                                    .Authenticate(_context.Account, _context.Environment, _context.Tenant.Id, null, ShowDialog.Never, null, _context.Environment.GetTokenAudience(AzureEnvironment.Endpoint.ResourceManager))
+                                    .AccessToken;
+            return GetClient()
+                    .RefreshTokens
+                    .GetFromExchangeAsync(grantType: _grantType, service: _endPoint, accessToken: armAccessToken)
+                    .GetAwaiter()
+                    .GetResult()
+                    .RefreshTokenProperty;
+        }
+
+        public PSRepositoryAttribute GetRepository(string repository)
+        {
+            return new ContainerRegistryRepositoryGetOperation(this, repository).ProcessRequest();
+        }
+
+        public IList<string> ListRepository(int? first)
+        {
+            return new ContainerRegistryRepositoryListOperation(this, first).ProcessRequest();
+        }
+
+        public PSDeletedRepository RemoveRepository(string repository)
+        {
+            return new ContainerRegistryRepositoryRemoveOperation(this, repository).ProcessRequest();
+        }
+
+        public PSRepositoryAttribute UpdateRepository(string repository, PSChangeableAttribute attribute)
+        {
+            new ContainerRegistryRepositoryUpdateOperation(this, repository, attribute).ProcessRequest();
+            return GetRepository(repository);
+        }
+
+        public PSAcrManifest ListManifest(string repository)
+        {
+            return new ContainerRegistryManifestListOperation(this, repository).ProcessRequest();
+        }
+
+        public PSManifestAttribute GetManifest(string repository, string manifest)
+        {
+            return new ContainerRegistryManifestGetOperation(this, repository, manifest).ProcessRequest();
+        }
+
+        public PSManifestAttribute UpdateManifest(string repository, string manifest, PSChangeableAttribute attribute)
+        {
+            new ContainerRegistryManifestUpdateOperation(this, repository, manifest, attribute).ProcessRequest();
+            return GetManifest(repository, manifest);
+        }
+
+        public PSManifestAttribute UpdateManifestByTag(string repository, string tag, PSChangeableAttribute attribute)
+        {
+            PSTagAttribute tagAttribute = GetTag(repository, tag);
+            return UpdateManifest(repository, tagAttribute.Attributes.Digest, attribute);
+        }
+
+        public bool RemoveManifest(string repository, string manifest)
+        {
+            return new ContainerRegistryManifestRemoveOperation(this, repository, manifest).ProcessRequest();
+        }
+
+        public bool RemoveManifestByTag(string repository, string tag)
+        {
+            PSTagAttribute tagAttribute = GetTag(repository, tag);
+            return RemoveManifest(repository, tagAttribute.Attributes.Digest);
+        }
+
+        public PSTagAttribute GetTag(string repository, string tag)
+        {
+            return new ContainerRegistryTagGetOperation(this, repository, tag).ProcessRequest();
+        }
+
+        public PSTagList ListTag(string repository)
+        {
+            return new ContainerRegistryTagListOperation(this, repository).ProcessRequest();
+        }
+
+        public bool RemoveTag(string repository, string tag)
+        {
+            return new ContainerRegistryTagRemoveOperation(this, repository, tag).ProcessRequest();
+        }
+
+        public PSTagAttribute UpdateTag(string repository, string tag, PSChangeableAttribute attribute)
+        {
+            new ContainerRegistryTagUpdateOperation(this, repository, tag, attribute).ProcessRequest();
+            return GetTag(repository, tag);
         }
     }
 }