+++ src/Accounts/Accounts/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated `Clear-AzContext` to clean token cache created by Az.ContainerRegistry
 * Shown correct client request id on debug message [#13745]
 * Added common Azure PowerShell exception type
 * Supported storage API 2019-06-01